backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.43
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
pydantic[email]==2.11.7
alembic==1.13.3
psycopg2-binary==2.9.5
python-dotenv==1.0.0
-